Updated Packaging and Added player associations.
