--- src/main/java/me/ironexception/universalfriends/IFriend.java
@@ -1,10 +1,0 @@
-package me.ironexception.universalfriends;
-
-import java.util.UUID;
-
-public interface IFriend {
-
-    String getName();
-    UUID getId();
-
-}
